delete specific data

+++ tools/importDeleteData.js
@@ -8,6 +8,8 @@ mongoose
 	.connect(process.env.DB_LOCAL)
 	.then(() => console.log('DATABASE CONNECTED'));
 
+var Food, User, Order, Review, Branch;
+
 if(process.argv[2] === '--import'){
 	importData();
 }
@@ -24,11 +26,7 @@ else{
 async function importData(){
 	try{
 		if(!process.argv[3]){
-			const Food = require('../models/FoodModel');
-			const Order = require('../models/OrderModel');
-			const User = require('../models/UserModel');
-			const Review = require('../models/ReviewModel');
-			const Branch = required('../models/BranchModel');
+			requireAllJSON();
 
 			const foods = JSON.parse(fs.readFileSync(`${__dirname}/data/foods.json`, 'utf-8'));
 			const users = JSON.parse(fs.readFileSync(`${__dirname}/data/users.json`, 'utf-8'));
@@ -47,13 +45,13 @@ async function importData(){
 			await Branch.create(branches);
 			console.log('Branch Data Imported Successfully');
 
-			return;
+			process.exit();
 		}
 
 		switch(process.argv[3]){
 			case 'F':
 			case 'f':
-			case 'food':
+			case 'foods':
 				const Food = require('../models/FoodModel');
 				const foods = JSON.parse(fs.readFileSync(`${__dirname}/data/foods.json`, 'utf-8'));
 				await Food.create(foods);
@@ -62,7 +60,7 @@ async function importData(){
 
 			case 'O':
 			case 'o':
-			case 'order':
+			case 'orders':
 				const Order = require('../models/OrderModel');
 				const orders = JSON.parse(fs.readFileSync(`${__dirname}/data/orders.json`, 'utf-8'));
 				await Order.create(orders);
@@ -97,7 +95,7 @@ async function importData(){
 				break;
 
 			default:
-				throw "Unrecognized argument";
+				throw 'Unrecognized argument';
 		}
 	}
 	catch(err){
@@ -109,23 +107,75 @@ async function importData(){
 }
 
 async function deleteData(){
-	const Food = require('../models/FoodModel');
-	const Order = require('../models/OrderModel');
-	const User = require('../models/UserModel');
-	const Review = require('../models/ReviewModel');
-	const Branch = require('../models/BranchModel');
-
 	try{
-		await Food.deleteMany();
-		await User.deleteMany();
-		await Order.deleteMany();
-		await Review.deleteMany();
-		await Branch.deleteMany();
+		if(!process.argv[3]){
+			requireAllJSON();
 
-		console.log('Data Deleted Successfully');
+			await Food.deleteMany();
+			await User.deleteMany();
+			await Order.deleteMany();
+			await Review.deleteMany();
+			await Branch.deleteMany();
+
+			console.log('Data Deleted Successfully');
+			process.exit();
+		}
+
+		switch(process.argv[3]){
+			case 'F':
+			case 'f':
+			case 'foods':
+				const Food = require('../models/FoodModel');
+				await Food.deleteMany();
+				console.log("Deleted Food data successfully.");
+				break;
+
+			case 'U':
+			case 'u':
+			case 'users':
+				const User = require('../models/UserModel');
+				await User.deleteMany();
+				console.log("Deleted User data successfully.");
+				break;
+
+			case 'O':
+			case 'o':
+			case 'orders':
+				const Order = require('../models/OrderModel');
+				await Order.deleteMany();
+				console.log("Deleted Order data successfully.");
+				break;
+
+			case 'R':
+			case 'r':
+			case 'reviews':
+				const Review = require('../models/ReviewModel');
+				await Review.deleteMany();
+				console.log("Deleted Review data successfully.");
+				break;
+
+			case 'B':
+			case 'b':
+			case 'branches':
+				const Branch = require('../models/BranchModel');
+				await Branch.deleteMany();
+				console.log("Deleted Branch data successfully.");
+				break;
+
+			default:
+				throw 'Unrecognized argument';
+		}
 	}
 	catch(err){
 		console.log(err);
 	}
 	process.exit(); // aggressively ends the process; if you put this outside on the main block, database will not be connected when Model is used (don't know why)
+}
+
+function requireAllJSON(){
+	Food = require('../models/FoodModel');
+	Order = require('../models/OrderModel');
+	User = require('../models/UserModel');
+	Review = require('../models/ReviewModel');
+	Branch = require('../models/BranchModel');
 }
